requirements.txt
pydantic
llama-index
llama-index-llms-gemini
python-dotenv
google-generativeai
